do not initialize jsPsych before
A common problem in JsPsych/javascript is to organize things so that it doesn't depend on variables that are not yet loaded. That was a problem for @AnafNeves paradigm of depending on some answer to filter out the correct stimuli. The solution to that is typically to define things inside functions, that are only called later in the script (and not at script initialization time) @BenjaminTribe